requirements.txt
# Discord Bot Dependencies
discord.py==2.5.0
aiohttp==3.8.5
aiohappyeyeballs==2.4.0